refactoring: Cars.js -> AllCars.js
